+++ apps/service/src/components/auth/index.ts
@@ -1,1 +1,2 @@
 export { default as LoginButton } from './LoginButton';
+export { default as KakaoCallback } from './KakaoCallback';
