+++ apps/service/src/constant/routes.ts
@@ -13,4 +13,5 @@ export const ROUTES = {
     const params = new URLSearchParams({ src: query.trim() });
     return `/watch?${params.toString()}`;
   },
+  LOGIN: '/login',
 } as const;
